+++ InvenTree/plugin/builtin/integration/test_mixins.py
@@ -1,26 +1,15 @@
-""" Unit tests for integration plugins """
-
-from django.test import TestCase
+"""Unit tests for plugin mixins"""
 from django.conf import settings
 from django.conf.urls import url, include
 from django.contrib.auth import get_user_model
 
-from datetime import datetime
-
 from plugin import IntegrationPluginBase
 from plugin.mixins import AppMixin, GlobalSettingsMixin, UrlsMixin, NavigationMixin
 from plugin.urls import PLUGIN_BASE
+from plugin.tests.test_integration import BaseMixinDefinition
 
 
-class BaseMixinDefinition:
-    def test_mixin_name(self):
-        # mixin name
-        self.assertEqual(self.mixin.registered_mixins[0]['key'], self.MIXIN_NAME)
-        # human name
-        self.assertEqual(self.mixin.registered_mixins[0]['human_name'], self.MIXIN_HUMAN_NAME)
-
-
-class GlobalSettingsMixinTest(BaseMixinDefinition, TestCase):
+class GlobalSettingsMixinTest(BaseMixinDefinition):
     MIXIN_HUMAN_NAME = 'Global settings'
     MIXIN_NAME = 'globalsettings'
     MIXIN_ENABLE_CHECK = 'has_globalsettings'
@@ -63,7 +52,7 @@ class GlobalSettingsMixinTest(BaseMixinDefinition, TestCase):
         self.assertEqual(self.mixin_nothing.get_globalsetting(''), '')
 
 
-class UrlsMixinTest(BaseMixinDefinition, TestCase):
+class UrlsMixinTest(BaseMixinDefinition):
     MIXIN_HUMAN_NAME = 'URLs'
     MIXIN_NAME = 'urls'
     MIXIN_ENABLE_CHECK = 'has_urls'
@@ -102,7 +91,7 @@ class UrlsMixinTest(BaseMixinDefinition, TestCase):
         self.assertEqual(self.mixin.internal_name, f'plugin:{self.mixin.slug}:')
 
 
-class AppMixinTest(BaseMixinDefinition, TestCase):
+class AppMixinTest(BaseMixinDefinition):
     MIXIN_HUMAN_NAME = 'App registration'
     MIXIN_NAME = 'app'
     MIXIN_ENABLE_CHECK = 'has_app'
@@ -117,7 +106,7 @@ class AppMixinTest(BaseMixinDefinition, TestCase):
         self.assertIn('plugin.samples.integration', settings.INSTALLED_APPS)
 
 
-class NavigationMixinTest(BaseMixinDefinition, TestCase):
+class NavigationMixinTest(BaseMixinDefinition):
     MIXIN_HUMAN_NAME = 'Navigation Links'
     MIXIN_NAME = 'navigation'
     MIXIN_ENABLE_CHECK = 'has_naviation'
@@ -146,71 +135,3 @@ class NavigationMixinTest(BaseMixinDefinition, TestCase):
         # navigation name
         self.assertEqual(self.mixin.navigation_name, 'abcd1')
         self.assertEqual(self.nothing_mixin.navigation_name, '')
-
-
-class IntegrationPluginBaseTests(TestCase):
-    """ Tests for IntegrationPluginBase """
-
-    def setUp(self):
-        self.plugin = IntegrationPluginBase()
-
-        class SimpeIntegrationPluginBase(IntegrationPluginBase):
-            PLUGIN_NAME = 'SimplePlugin'
-
-        self.plugin_simple = SimpeIntegrationPluginBase()
-
-        class NameIntegrationPluginBase(IntegrationPluginBase):
-            PLUGIN_NAME = 'Aplugin'
-            PLUGIN_SLUG = 'a'
-            PLUGIN_TITLE = 'a titel'
-            PUBLISH_DATE = "1111-11-11"
-            AUTHOR = 'AA BB'
-            DESCRIPTION = 'A description'
-            VERSION = '1.2.3a'
-            WEBSITE = 'http://aa.bb/cc'
-            LICENSE = 'MIT'
-
-        self.plugin_name = NameIntegrationPluginBase()
-
-    def test_action_name(self):
-        """check the name definition possibilities"""
-        # plugin_name
-        self.assertEqual(self.plugin.plugin_name(), '')
-        self.assertEqual(self.plugin_simple.plugin_name(), 'SimplePlugin')
-        self.assertEqual(self.plugin_name.plugin_name(), 'Aplugin')
-
-        # slug
-        self.assertEqual(self.plugin.slug, '')
-        self.assertEqual(self.plugin_simple.slug, 'simpleplugin')
-        self.assertEqual(self.plugin_name.slug, 'a')
-
-        # human_name
-        self.assertEqual(self.plugin.human_name, '')
-        self.assertEqual(self.plugin_simple.human_name, 'SimplePlugin')
-        self.assertEqual(self.plugin_name.human_name, 'a titel')
-
-        # description
-        self.assertEqual(self.plugin.description, '')
-        self.assertEqual(self.plugin_simple.description, 'SimplePlugin')
-        self.assertEqual(self.plugin_name.description, 'A description')
-
-        # author
-        self.assertEqual(self.plugin_name.author, 'AA BB')
-
-        # pub_date
-        self.assertEqual(self.plugin_name.pub_date, datetime(1111, 11, 11, 0, 0))
-
-        # version
-        self.assertEqual(self.plugin.version, None)
-        self.assertEqual(self.plugin_simple.version, None)
-        self.assertEqual(self.plugin_name.version, '1.2.3a')
-
-        # website
-        self.assertEqual(self.plugin.website, None)
-        self.assertEqual(self.plugin_simple.website, None)
-        self.assertEqual(self.plugin_name.website, 'http://aa.bb/cc')
-
-        # license
-        self.assertEqual(self.plugin.license, None)
-        self.assertEqual(self.plugin_simple.license, None)
-        self.assertEqual(self.plugin_name.license, 'MIT')
