+++ InvenTree/plugin/tests/__init__.py
@@ -1,0 +1,4 @@
+from .test_action import *
+from .test_api import *
+from .test_integration import *
+from .test_plugin import *
