+++ InvenTree/plugin/tests/test_plugin.py
